feat(analytics): switch plugin from gtag to gtm bootstrap
GA4 (gtag.js) を直接読み込んでいた client plugin を、GTM コンテナ
ブートストラップに置き換える。SPA ページビュー送信は GTM の
History Change Trigger に寄せ、plugin 側の router フックは廃止する。

runtimeConfig.public のキー名を gtagId から gtmId へ、対応する
環境変数も NUXT_PUBLIC_GTAG_ID から NUXT_PUBLIC_GTM_ID へ移行する。
未設定時の空文字フォールバックと plugins/gtm.client.ts 側の
fail-closed 判定は GA4 時代と同じ契約を維持する。

git mv 後に内容を全面置換しているため、git の rename 検知が
外れる点は許容する。
